pr-code-review-automation/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-code-review-automation"
version = "0.1.0"
description = "Bitbucket Server PR checkout, diff reconciliation, Maven compile, SpotBugs/Semgrep/PMD/Sonar scanner orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
dependencies = [
  "httpx>=0.27",
  "pyyaml>=6.0",
  "typer>=0.12.3",
  "pydantic>=2.6",
  "rich>=13.7",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3"]

[project.scripts]
pr-review = "pr_review.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pr_review = ["reports/templates/*.j2"]
